Update dependencies and refactor code for improved functionality
- Bump astropy version from 6.1.7 to 7.1.0 and adjust Python version requirement to >=3.11.
- Update astropy-iers-data, numpy, packaging, and PyYAML dependencies to newer versions.
- Refactor astrometry.py to suppress specific warnings and update template shifting logic.
- Modify fit.py to enhance logging, adjust GP length scale, and change scene merging behavior.
- Update pipeline.py to ensure images are converted to float32 after upsampling.
- Remove outdated make_testdata_old function from utils.py and adjust make_testdata for clarity.
- Improve scene summary logging in fit.py and ensure proper handling of low norm templates.

+++ src/mophongo/pipeline.py
@@ -179,7 +179,6 @@ class Pipeline:
         SparseFitter
             The fitter instance used for the final fit.
         """
-        from .psf import PSF
         from .fit import SparseFitter
         from .astro_fit import GlobalAstroFitter
         from .astrometry import AstroCorrect
@@ -240,9 +239,11 @@ class Pipeline:
             if k > 1:
                 if config.multi_resolution_method == "upsample":
                     print(f"upsampling image {idx} by factor {k}")
-                    images[idx] = block_replicate(images[idx], k, conserve_sum=True)
+                    images[idx] = block_replicate(images[idx], k, conserve_sum=True).astype(
+                        np.float32
+                    )
                     if weights_i is not None:
-                        weights_i = block_replicate(weights[idx], k) * k**2
+                        weights_i = block_replicate(weights[idx], k).astype(np.float32) * k**2
                     if wcs is not None:
                         wcs[idx] = wcs[0]
                 else:
